altum-bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.5.2
httpx==0.27.2
anthropic==0.40.0
google-generativeai==0.8.3
gspread==6.1.3
google-auth==2.35.0
python-dotenv==1.0.1
